requirements.txt
django
django-bootstrap-icons
django-staff-required
Pillow
django-debug-toolbar
black
